+++ content/document.js
@@ -18,6 +18,7 @@ const {
 const { getPopularities } = require("./popularities");
 const { getWikiHistories } = require("./wikihistories");
 const { getGitHistories } = require("./githistories");
+const { childrenFoldersForPath } = require("./document-paths");
 
 const {
   buildURL,
@@ -29,85 +30,6 @@ const {
   MEMOIZE_INVALIDATE,
 } = require("./utils");
 const Redirect = require("./redirect");
-
-function allDocumentPathsAsTree(root) {
-  const api = new fdir()
-    .withErrors()
-    .withRelativePaths()
-    .filter((filePath) => {
-      return (
-        filePath.endsWith(HTML_FILENAME) || filePath.endsWith(MARKDOWN_FILENAME)
-      );
-    })
-    .crawl(root);
-
-  const allPaths = [...api.sync()].map((p) => p.split(path.sep));
-  let tree = {};
-  for (const paths of allPaths) {
-    tree = addToTree(tree, paths);
-  }
-
-  return tree;
-}
-function addToTree(tree = {}, [current, ...rest]) {
-  if (rest.length === 0) {
-    tree[current] = null;
-    return tree;
-  }
-  tree[current] = addToTree(tree[current], rest);
-  return tree;
-}
-
-function flattenSubTree(t, prefix = [], recurse = false) {
-  const directChildren = [...Object.keys(t)]
-    .filter(
-      (k) =>
-        ![HTML_FILENAME, MARKDOWN_FILENAME].includes(k) &&
-        (t[k][HTML_FILENAME] === null || t[k][MARKDOWN_FILENAME] === null)
-    )
-    .map((k) =>
-      [
-        ...prefix,
-        k,
-        t[k][MARKDOWN_FILENAME] === null ? MARKDOWN_FILENAME : HTML_FILENAME,
-      ].join("/")
-    );
-  if (!recurse) {
-    return directChildren;
-  }
-
-  const recursiveChildren = [...Object.entries(t)]
-    .filter(([, v]) => v !== null)
-    .flatMap(([k, v]) => flattenSubTree(v, [...prefix, k], recurse));
-  return [...directChildren, ...recursiveChildren];
-}
-
-function childrenForPath(
-  tree,
-  [current, ...path],
-  recurse = false,
-  prefix = []
-) {
-  if (!current) {
-    return flattenSubTree(tree, prefix, recurse);
-  }
-  const parent = tree[current];
-  if (!parent) {
-    return [];
-  }
-  return childrenForPath(parent, path, recurse, [...prefix, current]);
-}
-
-// When running a production build we can afford to look up all files upfront.
-const ALL_DOCUMENT_PATHS_TREE =
-  process.env.NODE_ENV !== "production"
-    ? {}
-    : {
-        [CONTENT_ROOT]: allDocumentPathsAsTree(CONTENT_ROOT),
-        [CONTENT_TRANSLATED_ROOT]: CONTENT_TRANSLATED_ROOT
-          ? allDocumentPathsAsTree(CONTENT_TRANSLATED_ROOT)
-          : {},
-      };
 
 function buildPath(localeFolder, slug) {
   return path.join(localeFolder, slugToFolder(slug));
@@ -580,36 +502,8 @@ function findChildren(url, recursive = false) {
   const root = getRoot(locale);
   const folder = urlToFolderPath(url);
 
-  const base = path.join(root, folder);
-  const baseHTML = path.join(base, HTML_FILENAME);
-  const baseMarkdown = path.join(base, MARKDOWN_FILENAME);
-  // When running a production build we use our lookup.
-  if (process.env.NODE_ENV === "production") {
-    const childPaths = childrenForPath(
-      ALL_DOCUMENT_PATHS_TREE[root],
-      folder.split(path.sep),
-      recursive
-    );
-    return childPaths
-      .map((childFilePath) => path.dirname(childFilePath))
-      .map((folder) => read(folder));
-  } else {
-    const api = new fdir()
-      .withFullPaths()
-      .withErrors()
-      .filter((filePath) => {
-        return (
-          (filePath.endsWith(HTML_FILENAME) && !(filePath === baseHTML)) ||
-          (filePath.endsWith(MARKDOWN_FILENAME) && !(filePath === baseMarkdown))
-        );
-      })
-      .withMaxDepth(recursive ? Infinity : 1)
-      .crawl(base);
-    const childPaths = [...api.sync()];
-    return childPaths
-      .map((childFilePath) => path.relative(root, path.dirname(childFilePath)))
-      .map((folder) => read(folder));
-  }
+  const childPaths = childrenFoldersForPath(root, folder, recursive);
+  return childPaths.map((folder) => read(folder));
 }
 
 function move(oldSlug, newSlug, locale, { dry = false } = {}) {
